Correct header frame id, parameterized cloud topic
I parameterized the source cloud topic, since it can be different from "/camera/depth_registered/points" based on the user's setup. Secondly the header frame id for the bounding box marker as well as the selected point cloud should be the same as the source point cloud's, in order for it to appear in the correct position in rviz.